+++ awscli/customizations/s3/subcommands.py
@@ -115,7 +115,10 @@ INCLUDE = {'name': 'include', 'action': AppendFilter, 'nargs': 1,
            'dest': 'filters',
            'help_text': (
                "Don't exclude files or objects "
-               "in the command that match the specified pattern")}
+               "in the command that match the specified pattern. "
+               'See <a href="http://docs.aws.amazon.com/cli/latest/reference'
+               '/s3/index.html#use-of-exclude-and-include-filters">Use of '
+               'Exclude and Include Filters</a> for details.')}
 
 
 ACL = {'name': 'acl',
@@ -298,6 +301,23 @@ EXPIRES = {
 }
 
 
+METADATA = {
+    'name': 'metadata', 'cli_type_name': 'map',
+    'schema': {
+        'type': 'map',
+        'key': {'type': 'string'},
+        'value': {'type': 'string'}
+    },
+    'help_text': (
+        "A map of metadata to store with the objects in S3. This will be "
+        "applied to every object which is part of this request. In a sync, this"
+        "means that files which haven't changed won't receive the new metadata."
+        " When copying between two s3 locations, the metadata-directive "
+        "argument will default to 'REPLACE' unless otherwise specified."
+    )
+}
+
+
 METADATA_DIRECTIVE = {
     'name': 'metadata-directive', 'choices': ['COPY', 'REPLACE'],
     'help_text': (
@@ -385,6 +405,7 @@ def get_client(session, region, endpoint_url, verify, config=None):
     return session.create_client('s3', region_name=region,
                                  endpoint_url=endpoint_url, verify=verify,
                                  config=config)
+
 
 class S3Command(BasicCommand):
     def _run_main(self, parsed_args, parsed_globals):
@@ -649,7 +670,7 @@ class CpCommand(S3TransferCommand):
             "or <S3Uri> <S3Uri>"
     ARG_TABLE = [{'name': 'paths', 'nargs': 2, 'positional_arg': True,
                   'synopsis': USAGE}] + TRANSFER_ARGS + \
-                [METADATA_DIRECTIVE, EXPECTED_SIZE, RECURSIVE]
+                [METADATA, METADATA_DIRECTIVE, EXPECTED_SIZE, RECURSIVE]
 
 
 class MvCommand(S3TransferCommand):
@@ -659,8 +680,8 @@ class MvCommand(S3TransferCommand):
     USAGE = "<LocalPath> <S3Uri> or <S3Uri> <LocalPath> " \
             "or <S3Uri> <S3Uri>"
     ARG_TABLE = [{'name': 'paths', 'nargs': 2, 'positional_arg': True,
-                  'synopsis': USAGE}] + TRANSFER_ARGS + [METADATA_DIRECTIVE,
-                                                         RECURSIVE]
+                  'synopsis': USAGE}] + TRANSFER_ARGS +\
+                [METADATA, METADATA_DIRECTIVE, RECURSIVE]
 
 class RmCommand(S3TransferCommand):
     NAME = 'rm'
@@ -680,7 +701,8 @@ class SyncCommand(S3TransferCommand):
     USAGE = "<LocalPath> <S3Uri> or <S3Uri> " \
             "<LocalPath> or <S3Uri> <S3Uri>"
     ARG_TABLE = [{'name': 'paths', 'nargs': 2, 'positional_arg': True,
-                  'synopsis': USAGE}] + TRANSFER_ARGS
+                  'synopsis': USAGE}] + TRANSFER_ARGS + \
+                [METADATA, METADATA_DIRECTIVE]
 
 
 class MbCommand(S3TransferCommand):