+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 CHANGELOG
 =========
 
+1.9.10
+======
+
+* feature:``aws config``: Added support for dedicated hosts.
+* feature:``aws s3``: Added support for custom metadata in cp, mv, and sync.
+
+
 1.9.9
 =====